Remove re-exports under old names
