+++ gui/src/page_ui.rs
@@ -1,12 +1,12 @@
 use crate::content::{ContentLoader, ContentProvider};
-use crate::richtext::structured_document::BlockType;
 use fltk::{enums::Color, window};
+use rutle::structured_document::BlockType;
 use std::any::Any;
 
 /// A minimal UI abstraction layer for a page editor/viewer.
 ///
 /// It unifies the interactions needed by main.rs so different
-/// implementations (MarkdownEditor, TextDisplay, StructuredRichDisplay)
+/// implementations (MarkdownEditor, TextDisplay, StructuredRichUI)
 /// can be swapped without changing app logic.
 pub trait PageUI: ContentProvider + ContentLoader + 'static {
     // Subscribe to content change notifications (debounced by the app).
